move controller-manager bootstrappers to `pkg`
